added firebase sdk for hosting
