+++ src/App.js
@@ -9,7 +9,9 @@ import { Home } from './Pages/Home';
 import { About } from './Pages/About';
 import { Contact } from './Pages/Contact';
 import { Login } from './Pages/Login';
-import Product  from './Pages/Product';
+// import Product  from './Pages/Product';
+import ProductPage from './Pages/Product';
+import Products from './components/Products';
 
 export default function App() {
   return (
@@ -18,8 +20,8 @@ export default function App() {
     <Routes>
       <Route path="/" element={<Home />} />
     <Route path='/about' element={<About />} />
-    <Route path='/product' element={<Product /> } />
-    <Route path="/products/:id" element={<Product />} />
+    <Route path='/product' element={<Products /> } />
+    <Route path="/products/:id" element={<ProductPage />} />
     <Route path='/contact' element={<Contact  />} />
 
     <Route path='/login' element={<Login /> } />
