+++ src/components/Footer.js
@@ -13,8 +13,8 @@ const user = {
     'https://images.unsplash.com/photo-1472099645785-5658abf4ff4e?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
 }
 const navigation = [
-  { name: 'Home', href: '/', current: true },
-  { name: 'About', href: '/about', current: false },
+  { name: 'Home', href: '/', current: false },
+  { name: 'About', href: '/about', current: true },
   { name: 'Products', href: 'product', current: false },
   { name: 'Contact', href: 'contact', current: false },
 ]
@@ -56,7 +56,7 @@ export default function Example() {
                             key={item.name}
                             to={item.href}
                             className={classNames(
-                              item.current
+                            item.current
                                 ? 'bg-gray-900 text-white'
                                 : 'text-gray-300 hover:bg-gray-700 hover:text-white',
                               'px-3 py-2 rounded-md text-sm font-medium'
